<fix> 로그인 할 때 DB에서 해당 user의 아이템 목록 불러오기 기능 버그 수정 json과 RoomDB의 필드를 정확히 매칭
