+++ app/src/main/java/com/example/week2/data/item/Item.kt
@@ -3,6 +3,7 @@ package com.example.week2.data.item
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.google.gson.annotations.SerializedName
 
 @Entity(tableName = "item_table")
 data class Item(
@@ -11,5 +12,6 @@ data class Item(
     @ColumnInfo(name = "category") val category: String,
     @ColumnInfo(name = "image_uri") val item_image_url: String?,
     @ColumnInfo(name = "price") val price: Int,
+    @SerializedName("is_purchased")
     @ColumnInfo(name = "is_purchased") val isPurchased: Boolean = false
 )
