+++ app/src/main/java/com/example/week2/ui/activity/MainActivity.kt
@@ -12,6 +12,7 @@ import com.example.week2.data.item.ItemRepository
 import com.kakao.sdk.user.UserApiClient
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.launch
 import retrofit2.Call
 import retrofit2.Callback
@@ -19,10 +20,13 @@ import retrofit2.Response
 
 class MainActivity : AppCompatActivity() {
     private lateinit var repository: ItemRepository
+    private lateinit var apiService: ApiService
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        apiService = RetrofitClient.getInstance().create(ApiService::class.java)
 
         val db = AppRoomDatabase.getDatabase(applicationContext, CoroutineScope(Dispatchers.IO))
         repository = ItemRepository(db.itemDao())
@@ -123,15 +127,12 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun fetchUserItems(loginId: String) {
-        val retrofit = RetrofitClient.getInstance()
-        val apiService = retrofit.create(ApiService::class.java)
-
         apiService.getUserItems(loginId).enqueue(object : Callback<List<Item>> {
             override fun onResponse(call: Call<List<Item>>, response: Response<List<Item>>) {
                 if (response.isSuccessful) {
                     val userItems = response.body()
                     if (userItems != null) {
-                        saveUserItemsToLocalDB(userItems)
+                        resetLocalDatabase(userItems)
                     }
                     val intent = Intent(this@MainActivity, HomePageActivity::class.java)
                     startActivity(intent)
@@ -147,10 +148,76 @@ class MainActivity : AppCompatActivity() {
         })
     }
 
-    private fun saveUserItemsToLocalDB(items: List<Item>) {
+    private fun resetLocalDatabase(items: List<Item>) {
         CoroutineScope(Dispatchers.IO).launch {
+            repository.deleteAll()
+            Log.d("ResetDB", "All items deleted from the local database")
             repository.insertAll(items)
             Log.d("StoreActivity", "User items have been saved to the local database")
+            fetchShopItems()
         }
+    }
+
+    private fun fetchShopItems() {
+        apiService.getShopItems().enqueue(object : Callback<List<ServerItem>> {
+            override fun onResponse(call: Call<List<ServerItem>>, response: Response<List<ServerItem>>) {
+                if (response.isSuccessful) {
+                    val items = response.body()
+                    items?.let {
+                        Log.d("ShopItems", "JSON에서 받아온 items: $it")
+                        updateLocalDatabase(it)
+                    }
+                }
+            }
+
+            override fun onFailure(call: Call<List<ServerItem>>, t: Throwable) {
+                Log.e("ShopItems", "실패!!", t)
+            }
+        })
+    }
+
+    private fun updateLocalDatabase(serverItems: List<ServerItem>) {
+        CoroutineScope(Dispatchers.IO).launch {
+            val items = serverItems.map { it.toItem() }
+            val existingItems = mutableListOf<Item>()
+            repository.allItems.collect {
+                existingItems.addAll(it)
+                Log.d("ShopItems", "Collected existing items: $existingItems")
+
+                val itemsToUpdate = mutableListOf<Item>()
+                val itemsToInsert = mutableListOf<Item>()
+
+                items.forEach() { newItem ->
+                    val existingItem = existingItems.find { it.id == newItem.id }
+                    if (existingItem != null) {
+                        itemsToUpdate.add(newItem.copy(isPurchased = existingItem.isPurchased))
+                    } else {
+                        itemsToInsert.add(newItem)
+                    }
+                }
+                if (itemsToUpdate.isNotEmpty()) {
+                    repository.updateAll(itemsToUpdate)
+                    Log.d("ShopItems", "Updated items: $itemsToUpdate")
+                }
+                if (itemsToInsert.isNotEmpty()) {
+                    repository.insertAll(itemsToInsert)
+                    Log.d("ShopItems", "Inserted items: $itemsToInsert")
+                }
+                repository.allItems.collect { allItems ->
+                    Log.d("ShopItems", "After Updating: $allItems")
+                }
+            }
+        }
+    }
+
+    private fun ServerItem.toItem(): Item {
+        return Item(
+            id = this.id,
+            name = this.name,
+            category = this.category,
+            item_image_url = this.item_image_url,
+            price = this.price,
+            isPurchased = false // 초기값 설정
+        )
     }
 }
