+++ app/src/main/java/com/example/week2/ui/activity/StoreActivity.kt
@@ -54,7 +54,6 @@ class StoreActivity : AppCompatActivity(), ItemListAdapter.OnItemClickListener {
 
         val db = AppRoomDatabase.getDatabase(applicationContext, CoroutineScope(Dispatchers.IO))
         repository = ItemRepository(db.itemDao())
-        fetchShopItems()
         observeViewModel()
     }
 
@@ -78,71 +77,11 @@ class StoreActivity : AppCompatActivity(), ItemListAdapter.OnItemClickListener {
         return true
     }
 
-    private fun ServerItem.toItem(): Item {
-        return Item(
-            id = this.id,
-            name = this.name,
-            category = this.category,
-            item_image_url = this.item_image_url,
-            price = this.price,
-            isPurchased = false // 초기값 설정
-        )
-    }
-
-    private fun fetchShopItems() {
-        apiService.getShopItems().enqueue(object : Callback<List<ServerItem>> {
-            override fun onResponse(call: Call<List<ServerItem>>, response: Response<List<ServerItem>>) {
-                if (response.isSuccessful) {
-                    val items = response.body()
-                    items?.let {
-                        Log.d("ShopItems", "JSON에서 받아온 items: $it")
-                        updateLocalDatabase(it)
-                    }
-                }
-            }
-
-            override fun onFailure(call: Call<List<ServerItem>>, t: Throwable) {
-                Log.e("ShopItems", "실패!!", t)
-            }
-        })
-    }
-
     private fun observeViewModel() {
         itemViewModel.allItems.observe(this) { items ->
-            items.let { adapter.submitList(it) }
-        }
-    }
-
-    private fun updateLocalDatabase(serverItems: List<ServerItem>) {
-        CoroutineScope(Dispatchers.IO).launch {
-            val items = serverItems.map { it.toItem() }
-            val existingItems = mutableListOf<Item>()
-            repository.allItems.collect {
-                existingItems.addAll(it)
-                Log.d("ShopItems", "Collected existing items: $existingItems")
-
-                val itemsToUpdate = mutableListOf<Item>()
-                val itemsToInsert = mutableListOf<Item>()
-
-                items.forEach() { newItem ->
-                    val existingItem = existingItems.find { it.id == newItem.id }
-                    if (existingItem != null) {
-                        itemsToUpdate.add(newItem.copy(isPurchased = existingItem.isPurchased))
-                    } else {
-                        itemsToInsert.add(newItem)
-                    }
-                }
-                if (itemsToUpdate.isNotEmpty()) {
-                    repository.updateAll(itemsToUpdate)
-                    Log.d("ShopItems", "Updated items: $itemsToUpdate")
-                }
-                if (itemsToInsert.isNotEmpty()) {
-                    repository.insertAll(itemsToInsert)
-                    Log.d("ShopItems", "Inserted items: $itemsToInsert")
-                }
-                repository.allItems.collect { allItems ->
-                    Log.d("ShopItems", "After Updating: $allItems")
-                }
+            items.let {
+                adapter.submitList(it)
+                Log.d("StoreActivity", "Fetched items from RoomDB: $it")
             }
         }
     }
